Added i18n and translations 🌐
In this update, I added internationalization (i18n) support to the project, allowing users to use different languages in the interface. I included .json files containing translations for English (en_US.json) and Brazilian Portuguese (pt_BR.json). As a result, the project's GUI can now be displayed in both English and Portuguese, with translations applied to both the standard interface and the EasyGUI. This improvement enhances the project's accessibility and opens doors to a more diverse audience.